Info Projects/HugFace/requirements.txt
torch>=2.2.1
transformers>=4.36.2
sentence-transformers>=2.2.2
chromadb>=0.4.22
python-dotenv>=1.0.0
openai>=1.12.0
tiktoken>=0.6.0
rich>=13.7.0
typer>=0.9.0
pydantic>=2.5.0
langchain>=0.1.0
pypdf>=3.17.0
python-docx>=1.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
smolagents>=0.1.0
streamlit>=1.29.0
numpy>=1.26.4
matlabengine>=23.2.1
accelerate>=0.27.0
bitsandbytes>=0.41.0
safetensors>=0.4.2 